Update Controllers for Identity

+++ DatingApp.API/Controllers/MessagesController.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Security.Claims;
 using System.Threading.Tasks;
@@ -8,19 +7,17 @@ using DatingApp.API.Data;
 using DatingApp.API.Dtos;
 using DatingApp.API.Helpers;
 using DatingApp.API.Models;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace DatingApp.API.Controllers
 {
     [ServiceFilter(typeof(LogUserActivity))]
-    [Authorize]
     [Route("api/users/{userId}/[controller]")]
     [ApiController]
     public class MessagesController : ControllerBase
     {
-        private readonly IDatingRepository repo;
         private readonly IMapper mapper;
+        private readonly IDatingRepository repo;
 
         public MessagesController(IDatingRepository repo, IMapper mapper)
         {
@@ -31,36 +28,27 @@ namespace DatingApp.API.Controllers
         [HttpGet("{id}", Name = "GetMessage")]
         public async Task<IActionResult> GetMessage(int userId, int id)
         {
-            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-            {
-                return Unauthorized();
-            }
+            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value)) return Unauthorized();
 
             var messageFromRepo = await repo.GetMessage(id);
 
-            if (messageFromRepo == null)
-            {
-                return NotFound();
-            }
+            if (messageFromRepo == null) return NotFound();
 
             return Ok(messageFromRepo);
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetMessagesForUser(int userId, 
-            [FromQuery]MessageParams messageParams)
+        public async Task<IActionResult> GetMessagesForUser(int userId,
+            [FromQuery] MessageParams messageParams)
         {
-            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-            {
-                return Unauthorized();
-            }
+            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value)) return Unauthorized();
 
             messageParams.UserId = userId;
 
             var messagesFromRepo = await repo.GetMessagesForUser(messageParams);
 
             var messages = mapper.Map<IEnumerable<MessageToReturnDto>>(messagesFromRepo);
-            
+
             Response.AddPagination(messagesFromRepo.CurrentPage, messagesFromRepo.PageSize,
                 messagesFromRepo.TotalCount, messagesFromRepo.TotalPages);
 
@@ -70,10 +58,7 @@ namespace DatingApp.API.Controllers
         [HttpGet("thread/{recipientId}")]
         public async Task<IActionResult> GetMessageThread(int userId, int recipientId)
         {
-            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-            {
-                return Unauthorized();
-            }
+            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value)) return Unauthorized();
 
             var messagesFromRepo = await repo.GetMessageThread(userId, recipientId);
 
@@ -86,23 +71,17 @@ namespace DatingApp.API.Controllers
         public async Task<IActionResult> CreateMessage(int userId, MessageForCreationDto messageForCreationDto)
         {
             var sender = await repo.GetUser(userId);
-            
-            if (sender.Id != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-            {
-                return Unauthorized();
-            }
+
+            if (sender.Id != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value)) return Unauthorized();
 
             messageForCreationDto.SenderId = userId;
 
             var recipient = await repo.GetUser(messageForCreationDto.RecipientId);
 
-            if (recipient == null)
-            {
-                return BadRequest("Could not find user");
-            }
+            if (recipient == null) return BadRequest("Could not find user");
 
             var message = mapper.Map<Message>(messageForCreationDto);
-            
+
             repo.Add(message);
 
             if (await repo.SaveAll())
@@ -110,57 +89,36 @@ namespace DatingApp.API.Controllers
                 var messageToReturn = mapper.Map<MessageToReturnDto>(message);
                 return CreatedAtRoute("GetMessage", new {id = message.Id}, messageToReturn);
             }
-            
+
             throw new Exception("Creating the message failed on save");
         }
-        
+
         [HttpPost("{id}")]
         public async Task<IActionResult> DeleteMessage(int id, int userId)
         {
-            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-            {
-                return Unauthorized();
-            }
+            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value)) return Unauthorized();
 
             var messageFromRepo = await repo.GetMessage(id);
 
-            if (messageFromRepo.SenderId == userId)
-            {
-                messageFromRepo.SenderDeleted = true;
-            }
+            if (messageFromRepo.SenderId == userId) messageFromRepo.SenderDeleted = true;
 
-            if (messageFromRepo.RecipientId == userId)
-            {
-                messageFromRepo.RecipientDeleted = true;
-            }
+            if (messageFromRepo.RecipientId == userId) messageFromRepo.RecipientDeleted = true;
 
-            if (messageFromRepo.SenderDeleted && messageFromRepo.RecipientDeleted)
-            {
-                repo.Delete(messageFromRepo);
-            }
-            
-            if (await repo.SaveAll())
-            {
-                return NoContent();
-            }
-            
+            if (messageFromRepo.SenderDeleted && messageFromRepo.RecipientDeleted) repo.Delete(messageFromRepo);
+
+            if (await repo.SaveAll()) return NoContent();
+
             throw new Exception("Error deleting the message");
         }
 
         [HttpPost("{id}/read")]
         public async Task<IActionResult> MarkMessageAsRead(int userId, int id)
         {
-            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value))
-            {
-                return Unauthorized();
-            }
+            if (userId != int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value)) return Unauthorized();
 
             var message = await repo.GetMessage(id);
 
-            if (message.RecipientId != userId)
-            {
-                return Unauthorized();
-            }
+            if (message.RecipientId != userId) return Unauthorized();
 
             message.IsRead = true;
             message.DateRead = DateTime.Now;
@@ -169,6 +127,5 @@ namespace DatingApp.API.Controllers
 
             return NoContent();
         }
-        
     }
 }
